configration migrations remove-migration and patron repository

+++ EventosMusicales.Entities/Concierto.cs
@@ -13,7 +13,7 @@ namespace EventosMusicales.Entities
         public string Title { get; set; } = default!;
         public string Description { get; set; } = default!;
         public string Place { get; set; } = default!;
-        public double UnitPrice { get; set; }
+        public Decimal UnitPrice { get; set; }
         public int GenreId { get; set; }
         public DateTime DateEvent { get; set; }
         public string? ImageUrl { get; set; }
